+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 -   Revert Card component removal #7167.
+-   Add rowKey prop to Table and TableCard component. #7196
 
 # 7.0.0